separate out collision system
