+++ izzet/systems/shared/shared.go
@@ -1,8 +1,12 @@
 package shared
 
-import "github.com/kkevinchou/izzet/izzet/entities"
+import (
+	"github.com/kkevinchou/izzet/izzet/entities"
+	"github.com/kkevinchou/kitolib/spatialpartition"
+)
 
 type GameWorld interface {
 	GetEntityByID(int) *entities.Entity
 	Entities() []*entities.Entity
+	SpatialPartition() *spatialpartition.SpatialPartition
 }
